Update openbitcoinatm.ino
Attempt at embedded QR encoding.
I took sparkfun's thermal printer library and merged it with this project.
Sparkfun used a novel ascii character block method for printing out a qr code, but it didn't work on my thermal printer, perhaps because the current draw was too intense for the full width codes it was printing out.
Then I tried printing out the bitmap directly, but the printer froze, think the buffer ran out of space?
So instead I came up with a crude way to write the qrcode to the SD card as an array, then pick it up again in printbitmap() and print it like the preloaded bitmaps in the unforked code. But now I'm just getting garbage back from Serial. I trimmed a lot of the int declarations to chars and shorts to free up memory but that didn't help. 
Ordered a 9V2A power supply, maybe I'll use that and go back to sparkfun's character printing method once it arrives in the mail.