Generates private keys / QR codes
I've swapped the printer and uno for a screen and the mega so I could test this repeatedly without worrying about paper. It also lightens the machine and speeds up the process a little.
I got the machine to generate private keys using sha-256 and a determined seed including the unixtime and some other fixed values. 
The same values would be used in a desktop application which uses the same algorithm to generate identical keys. That application would then instruct a wallet through RPC calls to send bitcoin to the addresses which go with the keys. The result is an offline machine which can be remotely 'loaded up' with bitcoin from an internet-connected computer.
In the future, the machine could support a GPS chip which would effectively scramble the keys if the machine were moved or stolen. I put in placeholder values for this feature.
The machine also generates QR codes on its own.
SD card no longer required, but you'll need a RTC clock for the time seeding. (I think you can just use the data logging shield's clock anyway). Idea: EEPROM could be used for a transaction log instead of an SD card.
It takes a lot more RAM for the crypto libraries, so this update requires a Mega instead of an Uno. Hopefully the program can be further optimized so that an Uno is enough.